requirement.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
absl-py=1.0.0=pypi_0
aiohttp=3.8.1=pypi_0
aiosignal=1.2.0=pypi_0
albumentations=1.1.0=pypi_0
appdirs=1.4.4=pypi_0
argon2-cffi=21.3.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
asttokens=2.0.5=pypi_0
async-timeout=4.0.1=pypi_0
attrs=21.2.0=pypi_0
audioread=2.1.9=pypi_0
autopep8=1.6.0=pyhd3eb1b0_0
backcall=0.2.0=pypi_0
black=22.1.0=pypi_0
blas=1.0=mkl
bleach=4.1.0=pypi_0
bzip2=1.0.8=h7b6447c_0
ca-certificates=2021.10.26=h06a4308_2
cachetools=4.2.4=pypi_0
certifi=2021.10.8=py38h06a4308_0
cffi=1.15.0=pypi_0
charset-normalizer=2.0.7=pypi_0
click=8.0.3=pypi_0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.11.0=pypi_0
datasets=1.17.0=pypi_0
debugpy=1.5.1=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
deprecated=1.2.13=pypi_0
dill=0.3.4=pypi_0
einops=0.3.2=pypi_0
entrypoints=0.4=pypi_0
executing=0.8.2=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.4.2=pypi_0
fonttools=4.28.2=pypi_0
freetype=2.11.0=h70c0345_0
frozenlist=1.2.0=pypi_0
fsspec=2021.11.0=pypi_0
future=0.18.2=pypi_0
giflib=5.2.1=h7b6447c_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
google-auth=2.3.3=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
grpcio=1.41.1=pypi_0
h5py=3.6.0=pypi_0
huggingface-hub=0.4.0=pypi_0
humanize=3.12.0=pypi_0
hyperpyyaml=1.0.0=pypi_0
idna=3.3=pypi_0
imageio=2.11.1=pypi_0
importlib-metadata=4.8.2=pypi_0
importlib-resources=5.4.0=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.9.1=pypi_0
ipython=8.0.1=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=7.6.5=pypi_0
jedi=0.18.1=pypi_0
jinja2=3.0.3=pypi_0
joblib=1.1.0=pypi_0
jpeg=9b=h024ee3a_2
jsonschema=4.4.0=pypi_0
jupyter-client=7.1.2=pypi_0
jupyter-core=4.9.2=pypi_0
jupyterlab-pygments=0.1.2=pypi_0
jupyterlab-widgets=1.0.2=pypi_0
kiwisolver=1.3.2=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgomp=9.3.0=h5101ec6_17
libiconv=1.15=h63c8f33_5
libidn2=2.3.2=h7f8727e_0
libpng=1.6.37=hbc83047_0
librosa=0.8.1=pypi_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuv=1.40.0=h7b6447c_0
libwebp=1.2.0=h89dd481_0
libwebp-base=1.2.0=h27cfd23_0
llvmlite=0.38.0=pypi_0
lz4-c=1.9.3=h295c915_1
markdown=3.3.5=pypi_0
markupsafe=2.1.0=pypi_0
matplotlib=3.5.0=pypi_0
matplotlib-inline=0.1.3=pypi_0
mistune=0.8.4=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
monai=0.8.0=pypi_0
multidict=5.2.0=pypi_0
multiprocess=0.70.12.2=pypi_0
mypy-extensions=0.4.3=pypi_0
nbclient=0.5.11=pypi_0
nbconvert=6.4.2=pypi_0
nbformat=5.1.3=pypi_0
ncurses=6.3=h7f8727e_2
nest-asyncio=1.5.4=pypi_0
nettle=3.7.3=hbbd107a_1
networkx=2.6.3=pypi_0
nibabel=3.2.1=pypi_0
ninja=1.10.2=py38hd09550d_3
notebook=6.4.8=pypi_0
numba=0.55.0=pypi_0
numexpr=2.8.1=pypi_0
numpy=1.21.2=py38h20f2e39_0
numpy-base=1.21.2=py38h79a1101_0
oauthlib=3.1.1=pypi_0
olefile=0.46=pyhd3eb1b0_0
opencv-python-headless=4.5.4.58=pypi_0
openh264=2.1.0=hd408876_0
openssl=1.1.1l=h7f8727e_0
packaging=21.2=pypi_0
pandas=1.3.5=pypi_0
pandocfilters=1.5.0=pypi_0
parso=0.8.3=pypi_0
pathspec=0.9.0=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=8.4.0=py38h5aabda8_0
pip=21.2.4=py38h06a4308_0
platformdirs=2.5.1=pypi_0
pooch=1.5.2=pypi_0
prometheus-client=0.13.1=pypi_0
prompt-toolkit=3.0.28=pypi_0
protobuf=3.19.1=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pyarrow=6.0.1=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycodestyle=2.8.0=pyhd3eb1b0_0
pycparser=2.21=pypi_0
pydeprecate=0.3.1=pypi_0
pygments=2.11.2=pypi_0
pyparsing=2.4.7=pypi_0
pyrsistent=0.18.1=pypi_0
python=3.8.12=h12debd9_0
python-dateutil=2.8.2=pypi_0
pytorch=1.10.1=py3.8_cuda11.3_cudnn8.2.0_0
pytorch-lightning=1.5.2=pypi_0
pytorch-mutex=1.0=cuda
pytz=2021.3=pypi_0
pywavelets=1.2.0=pypi_0
pyyaml=6.0=pypi_0
pyzmq=22.3.0=pypi_0
qudida=0.0.4=pypi_0
readline=8.1=h27cfd23_0
regex=2021.11.10=pypi_0
requests=2.26.0=pypi_0
requests-oauthlib=1.3.0=pypi_0
resampy=0.2.2=pypi_0
rsa=4.7.2=pypi_0
ruamel-yaml=0.17.20=pypi_0
ruamel-yaml-clib=0.2.6=pypi_0
sacremoses=0.0.47=pypi_0
scikit-image=0.18.3=pypi_0
scikit-learn=1.0.1=pypi_0
scipy=1.7.2=pypi_0
seaborn=0.11.2=pypi_0
send2trash=1.8.0=pypi_0
sentencepiece=0.1.96=pypi_0
setuptools=58.0.4=py38h06a4308_0
setuptools-scm=6.3.2=pypi_0
simpleitk=2.1.1=pypi_0
six=1.16.0=pyhd3eb1b0_0
soundfile=0.10.3.post1=pypi_0
speechbrain=0.5.11=pypi_0
sqlite=3.36.0=hc218d9a_0
stack-data=0.2.0=pypi_0
tables=3.7.0=pypi_0
tensorboard=2.7.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.0=pypi_0
terminado=0.13.1=pypi_0
testpath=0.5.0=pypi_0
textgrid=1.5=pypi_0
threadpoolctl=3.0.0=pypi_0
tifffile=2021.11.2=pypi_0
timm=0.4.12=pypi_0
tk=8.6.11=h1ccaba5_0
tokenizers=0.11.6=pypi_0
toml=0.10.2=pyhd3eb1b0_0
tomli=1.2.2=pypi_0
torchaudio=0.10.1=py38_cu113
torchio=0.18.68=pypi_0
torchmetrics=0.6.0=pypi_0
torchvision=0.11.2=py38_cu113
tornado=6.1=pypi_0
tqdm=4.62.3=pypi_0
traitlets=5.1.1=pypi_0
transformers=4.17.0=pypi_0
typing_extensions=3.10.0.2=pyh06a4308_0
urllib3=1.26.7=pypi_0
wcwidth=0.2.5=pypi_0
webencodings=0.5.1=pypi_0
werkzeug=2.0.2=pypi_0
wheel=0.37.0=pyhd3eb1b0_1
widgetsnbextension=3.5.2=pypi_0
wrapt=1.13.3=pypi_0
xxhash=2.0.2=pypi_0
xz=5.2.5=h7b6447c_0
yarl=1.7.2=pypi_0
zipp=3.6.0=pypi_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0
